chore: move components to dedicated dir
