+++ packages/cli/src/components/audit-licenses/audit-result.tsx
