+++ packages/cli/src/components/audit-licenses/license-list.tsx
