+++ packages/cli/src/components/audit-licenses/no-licenses-found-result.tsx
