+++ packages/cli/src/components/audit-licenses/success-result.tsx
