+++ packages/cli/src/components/audit-licenses/verbose-view/found-license-section.tsx
