+++ packages/cli/src/components/audit-licenses/verbose-view/not-found-license-section.tsx
