+++ packages/cli/src/components/audit-licenses/verbose-view/verbose-view.tsx
